% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c1v1.R
\name{c1v1}
\alias{c1v1}
\title{A Concentration Conversion Function}
\usage{
c1v1(c1 = 0, v1 = 0, c2 = 0)
}
\arguments{
\item{c1}{Target concentration in nM or ng/uL}

\item{v1}{Target volume in uL}

\item{c2}{Current concentration in nM or ng/uL}
}
\description{
A function to determine volume given target concentration. Will print volume of stock required to get to target concentration. Both concentrations must be in same units
}
\examples{
c1v1(50, 10, 300)
> 1.67 uL
}
\keyword{concentration,}
\keyword{dilution}
\keyword{target,}
\keyword{volume,}
