% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngul_conv.R
\name{nguL_conv}
\alias{nguL_conv}
\title{A Concentration Conversion Function}
\usage{
nguL_conv(nguL = 0, bp = 0)
}
\arguments{
\item{nguL}{the concentration in ng/uL}

\item{bp}{the length in base pairs}
}
\description{
A function to convert between ng/uL and nM (nanograms/uL and nanomolar). Will print the nM concentration. Requires base pair length.
}
\examples{
nguL_conv(30, 350)
> 129.87 nM
}
\keyword{concentration,}
\keyword{conversion,}
\keyword{ng/uL}
